+++ src/components/cards/cardPost/cardPost.styles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { TextMuted } from '../../ui/typography/TextMuted';
+import { TextMuted } from '../../ui-system/typography/TextMuted';
 
 export const SubHeader = styled(TextMuted)`
   font-size: var(--font-size-smaller);
