+++ src/components/cards/cardPost/cardPost.tsx
@@ -1,11 +1,11 @@
 import { PostMetaData } from '../../../types/FrontMatter';
-import { Card } from '../../ui/card/Card';
-import { Column } from '../../ui/container/Column';
-import { Row } from '../../ui/container/Row';
-import { UnstyledInternalLink } from '../../ui/content/UnstyledLink';
-import { MarginSmall } from '../../ui/margins/MarginSmall';
-import { Header } from '../../ui/typography/Header';
-import { TextMuted } from '../../ui/typography/TextMuted';
+import { Card } from '../../ui-system/card/Card';
+import { Column } from '../../ui-system/container/Column';
+import { Row } from '../../ui-system/container/Row';
+import { UnstyledInternalLink } from '../../ui-system/content/UnstyledLink';
+import { MarginSmall } from '../../ui-system/margins/MarginSmall';
+import { Header } from '../../ui-system/typography/Header';
+import { TextMuted } from '../../ui-system/typography/TextMuted';
 import { animationSpring } from '../animations';
 import { SubHeader } from './cardPost.styles';
