+++ src/components/cards/cardProject/cardProject.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Card } from '../../ui/card/Card';
-import { Column } from '../../ui/container/Column';
-import { Fill } from '../../ui/container/Fill';
-import { Row } from '../../ui/container/Row';
-import { UnstyledLink } from '../../ui/content/UnstyledLink';
-import { MarginSmall } from '../../ui/margins/MarginSmall';
-import { Header } from '../../ui/typography/Header';
+import { Card } from '../../ui-system/card/Card';
+import { Column } from '../../ui-system/container/Column';
+import { Fill } from '../../ui-system/container/Fill';
+import { Row } from '../../ui-system/container/Row';
+import { UnstyledLink } from '../../ui-system/content/UnstyledLink';
+import { MarginSmall } from '../../ui-system/margins/MarginSmall';
+import { Header } from '../../ui-system/typography/Header';
 import { animationSpring } from '../animations';
 
 interface CardProjectProps {
