+++ src/components/footer/footer.tsx
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
 import { site } from '../../data/site';
-import { Column } from '../ui/container/Column';
-import { Fill } from '../ui/container/Fill';
-import { ResponsiveContainer } from '../ui/container/ResponsiveContainer';
-import { Row } from '../ui/container/Row';
-import { IconGitHub } from '../ui/icons/IconGitHub';
-import { IconMail } from '../ui/icons/IconMail';
-import { MarginMedium } from '../ui/margins/MarginMedium';
-import { MarginSmall } from '../ui/margins/MarginSmall';
+import { Column } from '../ui-system/container/Column';
+import { Fill } from '../ui-system/container/Fill';
+import { ResponsiveContainer } from '../ui-system/container/ResponsiveContainer';
+import { Row } from '../ui-system/container/Row';
+import { IconGitHub } from '../ui-system/icons/IconGitHub';
+import { IconMail } from '../ui-system/icons/IconMail';
+import { MarginMedium } from '../ui-system/margins/MarginMedium';
+import { MarginSmall } from '../ui-system/margins/MarginSmall';
 
 const FooterText = styled.p`
   margin: 0;
