+++ src/components/ui-system/typography/TextCenter.tsx
