+++ src/layouts/blogPost.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { CardPost } from '../components/cards/cardPost/CardPost';
 import { SeoBlogPost } from '../components/seo/SeoBlogPost';
-import { Column } from '../components/ui/container/Column';
-import { ResponsiveContainer } from '../components/ui/container/ResponsiveContainer';
-import { ResponsiveGrid } from '../components/ui/container/ResponsiveGrid';
-import { Divider } from '../components/ui/content/Divider';
-import { MarginLarge } from '../components/ui/margins/MarginLarge';
-import { MarginMedium } from '../components/ui/margins/MarginMedium';
-import { TextMuted } from '../components/ui/typography/TextMuted';
+import { Column } from '../components/ui-system/container/Column';
+import { ResponsiveContainer } from '../components/ui-system/container/ResponsiveContainer';
+import { ResponsiveGrid } from '../components/ui-system/container/ResponsiveGrid';
+import { Divider } from '../components/ui-system/content/Divider';
+import { MarginLarge } from '../components/ui-system/margins/MarginLarge';
+import { MarginMedium } from '../components/ui-system/margins/MarginMedium';
+import { TextMuted } from '../components/ui-system/typography/TextMuted';
 import { usePrevAndNextBlogPost } from '../hooks/usePrevAndNextBlogPost';
 import { PostLayoutProps } from '../types/Layout';
 import { toPostMetaData } from '../utils/postMappings';
