+++ src/pages/index.tsx
@@ -3,15 +3,15 @@ import { SectionArticles } from '../components/sections/SectionArticles';
 import { SectionHero } from '../components/sections/SectionHero';
 import { SectionProjects } from '../components/sections/SectionProjects';
 import { ToolBox } from '../components/toolbox/ToolBox';
-import { ButtonPrimary } from '../components/ui/buttons/ButtonPrimary';
-import { Column } from '../components/ui/container/Column';
-import { Fill } from '../components/ui/container/Fill';
-import { ResponsiveContainer } from '../components/ui/container/ResponsiveContainer';
-import { Row } from '../components/ui/container/Row';
-import { UnstyledInternalLink } from '../components/ui/content/UnstyledLink';
-import { MarginLarge } from '../components/ui/margins/MarginLarge';
-import { MarginMedium } from '../components/ui/margins/MarginMedium';
-import { TextCenter } from '../components/ui/typography/TextCenter';
+import { ButtonPrimary } from '../components/ui-system/buttons/ButtonPrimary';
+import { Column } from '../components/ui-system/container/Column';
+import { Fill } from '../components/ui-system/container/Fill';
+import { ResponsiveContainer } from '../components/ui-system/container/ResponsiveContainer';
+import { Row } from '../components/ui-system/container/Row';
+import { UnstyledInternalLink } from '../components/ui-system/content/UnstyledLink';
+import { MarginLarge } from '../components/ui-system/margins/MarginLarge';
+import { MarginMedium } from '../components/ui-system/margins/MarginMedium';
+import { TextCenter } from '../components/ui-system/typography/TextCenter';
 import { useContent } from '../context/ContentContext';
 import { site } from '../data/site';
 import { toolboxCategories } from '../data/toolBoxCategories';
